+++ src/resources/filters/crossref/format.lua
@@ -191,7 +191,9 @@ function formatNumberOption(type, order, default)
     if section then
       tprepend(option, { pandoc.Str(tostring(section[1]) .. ".") })
     end
-    return option
+    quarto.utils.dump(styleRaw)
+    print("HERE!", option)
+    return { option }
   end
 
 end
